+++ src/titiler/extensions/titiler/extensions/__init__.py
@@ -1,6 +1,6 @@
 """titiler.extensions"""
 
-__version__ = "2.0.0dev0"
+__version__ = "2.0.0a1"
 
 from .cogeo import cogValidateExtension  # noqa
 from .render import stacRenderExtension  # noqa